+++ hdsr_fewspy/api_calls/get_parameters.py
@@ -49,7 +49,7 @@ class GetParameters(GetRequest):
         ]
 
     @property
-    def allowed_output_choices(self) -> List[str]:
+    def allowed_output_choices(self) -> List[OutputChoices]:
         return [
             OutputChoices.json_response_in_memory,
             OutputChoices.xml_response_in_memory,