+++ hdsr_fewspy/api_calls/get_timezone_id.py
@@ -27,7 +27,7 @@ class GetTimeZoneId(GetRequest):
         return []
 
     @property
-    def allowed_output_choices(self) -> List[str]:
+    def allowed_output_choices(self) -> List[OutputChoices]:
         return [
             OutputChoices.json_response_in_memory,
             OutputChoices.xml_response_in_memory,