+++ hdsr_fewspy/api_calls/time_series/get_time_series_single.py
@@ -30,7 +30,7 @@ class GetTimeSeriesSingle(GetTimeSeriesBase):
                 logger.warning(f"drop_missing_values is not used for output_choice {self.output_choice}")
 
     @property
-    def allowed_output_choices(self) -> List[str]:
+    def allowed_output_choices(self) -> List[OutputChoices]:
         return [
             OutputChoices.json_response_in_memory,
             OutputChoices.xml_response_in_memory,