+++ hdsr_fewspy/examples/point.py
@@ -20,6 +20,7 @@ def run_example_point():
     # raw = 20 csvs (alleen qg)
     # work = 18 csvs (alleen qb)
     # validated = 18 csvs (alleen qb)
+    setup_logging()
 
     logger.info("start run_example_point")
 
@@ -33,26 +34,6 @@ def run_example_point():
     df_all_locs = api.get_locations(output_choice=hdsr_fewspy.OutputChoices.pandas_dataframe_in_memory)
     df_kw_locs = df_all_locs[df_all_locs.index.str.startswith("KW")]
     locations = df_kw_locs.index.to_list()
-
-    # ik vermoed dat de laatste download sessie is onderbroken (pycharm was afgesloten..)
-    # ik heb geen logging file
-    # laaste loc in dir hdsr_fewspy_20230601_130806_validated/ = KW446911
-    # locations.index("KW446911")
-    # 442
-    # len(locations)
-    # 1143
-    # dus ergens op 1/3 is sessie onderbroken?
-    # locations = locations[442:]
-
-    # ik vermoed dat de laatste download sessie is onderbroken (pycharm was afgesloten..)
-    # ik heb geen logging file
-    # laaste loc in dir hdsr_fewspy_20230601_130806_validated/ = KW446911
-    # locations.index("KW761201")
-    # 1104
-    # len(locations)
-    # 1143
-    # dus ergens op 96% is sessie onderbroken?
-    locations = locations[1104:]
 
     # request data from Api
     api.get_time_series_multi(
@@ -76,7 +57,3 @@ def setup_logging() -> None:
     root_logger.addHandler(stream_handler)
     root_logger.setLevel(min([handler.level for handler in root_logger.handlers]))
     root_logger.info("setup logging done")
-
-
-setup_logging()
-run_example_point()