+++ HISTORY.rst
@@ -3,6 +3,7 @@
 - add logging to examples
 - add timeseries flag description (Deltares url) to readme
 - add __repr__ to PiSettings so that non-python users (R, etc.) can print those settings
+- TODO: updated default pi_settings to new datascience FEWS machine (production)
 
 1.11 (2023-6-1)
 ----------------